DeepLabV3Plus/requirements.txt
torch
torchvision~=0.17.0
numpy~=1.24.3
pillow~=9.4.0
scikit-learn~=1.3.0
tqdm~=4.65.0
matplotlib~=3.7.2
visdom
datasets~=2.12.0
pandas~=2.0.3
opencv-python~=4.9.0.80